+++ src/main/java/forestry/energy/tiles/BiogasEngineBlockEntity.java
@@ -11,7 +11,6 @@
 package forestry.energy.tiles;
 
 import javax.annotation.Nullable;
-import java.util.List;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -41,6 +40,7 @@ import forestry.api.fuels.FuelManager;
 import forestry.core.config.Constants;
 import forestry.core.fluids.FilteredTank;
 import forestry.core.fluids.FluidHelper;
+import forestry.core.fluids.FluidTagFilter;
 import forestry.core.fluids.StandardTank;
 import forestry.core.fluids.TankManager;
 import forestry.core.tiles.ILiquidTankTile;
@@ -68,7 +68,7 @@ public class BiogasEngineBlockEntity extends EngineBlockEntity implements Worldl
 		setInternalInventory(new InventoryEngineBiogas(this));
 
 		this.fuelTank = new FilteredTank(Constants.ENGINE_TANK_CAPACITY).setFilters(FuelManager.biogasEngineFuel.keySet());
-		this.heatingTank = new FilteredTank(Constants.ENGINE_TANK_CAPACITY, true, false).setFilters(List.of(Fluids.LAVA));
+		this.heatingTank = new FilteredTank(Constants.ENGINE_TANK_CAPACITY, true, false).setFilter(FluidTagFilter.LAVA);
 		this.burnTank = new StandardTank(BUCKET_VOLUME, false, false);
 
 		this.tankManager = new TankManager(this, fuelTank, heatingTank, burnTank);